pyproject.toml
[project]
name = "falk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Josep Maria Sanz Xicola", email = "jsanz83@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.46",
    "tinytuya>=1.17.4",
]

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[tool.setuptools.dynamic]
version = {attr = 'falk._version.__version__'}
readme = {file = ['README.md']}

